+++ src/clj/withings_client/tokens.clj
@@ -63,9 +63,9 @@
 ;; client_secret=d9286311451fc6ed71b372a075c58c5058be158f56a77865e43ab3783255424f&
 ;; refresh_token=9697c3d06ccfd1ca302f5a527d345a9f99ea88a2"
 ;; 'https://wbsapi.withings.net/v2/oauth2'
-;;
 
-(defn refresh-token
+(defn refresh
+  "when errors, returns {}"
   [{:keys [cid secret refresh]}]
   (-> (hc/post
        oauth2-uri
@@ -74,19 +74,27 @@
                        :grant_type    "refresh_token"
                        :client_id     cid
                        :client_secret secret
-                       :refresh_token refresh}})))
+                       :refresh_token refresh}})
+      (get-in [:body :body])))
 
-;; {
-;;   "status": 0,
-;;   "body": {
-;;            "userid": "363",
-;;            "access_token": "a075f8c14fb8df40b08ebc8508533dc332a6910a",
-;;            "refresh_token": "f631236f02b991810feb774765b6ae8e6c6839ca",
-;;            "expires_in": 10800,
-;;            "scope": "user.info,user.metrics",
-;;            "csrf_token": "PACnnxwHTaBQOzF7bQqwFUUotIuvtzSM",
-;;            "token_type": "Bearer"}}
+;; (refresh {:cid "f7164783bfc573217510d38c07176b798daa2a9e78edf1e320e6c1f0e5a5fa35"
+;;           :secret "027a7105dfc049d9fe4722ab8f6dbdc66a3931702b57976f437af2368813e3de"
+;;           :refresh "160395aa1d38b6635587ab4e30c410f5a6c86044"})
 
-(defn update-tokens-by-userid
- [params]
- (users/update-tokens-by-userid! params))
+;; {:userid 13662696,
+;;  :access_token "71781b62732d2e95ec4d2a349735cdb8b993b52d",
+;;  :refresh_token "62cfc36a72820d50a6d83c3857849e978ff007b5",
+;;  :scope "user.metrics,user.activity,user.info",
+;;  :expires_in 10800,
+;;  :token_type "Bearer"}
+
+(defn restore!
+  "params には userid, access, refresh, access, "
+  [params]
+  (if (empty? params)
+   {:error "empty"}
+   (users/update-tokens-by-userid! params)))
+
+(defn refresh-and-restore!
+  [params]
+  (-> params refresh restore!))